tomcat/src/main/java/com/techcourse/db/InMemoryUserRepository.java
package com.techcourse.db;

import com.techcourse.model.User;
import java.util.Map;
import java.util.Optional;
import java.util.concurrent.ConcurrentHashMap;
import java.util.concurrent.atomic.AtomicLong;

public class InMemoryUserRepository {

    private static final Map<String, User> database = new ConcurrentHashMap<>();
    private static final AtomicLong sequence = new AtomicLong(1);

    static {
        reset();
    }

    public static void reset() {
        database.clear();
        sequence.set(1);

        User user = new User(sequence.getAndIncrement(), "gugu", "password", "hkkang@woowahan.com");
        database.put(user.getAccount(), user);
    }

    public static void save(User user) {
        User newUser = new User(sequence.getAndIncrement(), user);
        database.put(newUser.getAccount(), newUser);
    }

    public static Optional<User> findByAccount(String account) {
        return Optional.ofNullable(database.get(account));
    }

    private InMemoryUserRepository() {
    }
}
